Position light relative to eye space
And do the modelview transform to translate the light vector properly
in to object space.

Because model and view matrix are combined in opengl, this means we
can't ever change the view matrix without it breaking. But at least
then we can just move the light relative to the view transform.