Unify coordinate systems and physics for choper and missiles
Map tiles were not being placed in the expected world coordinates,
they were on an inverted coordinate system with 0,0 in bottom right
instead of top left. Fixed this by moving the heli position in
positive coordinate system and translating world to inverse of the
heli coordinates.

Also optimised the map renderer by using the mins/maxs of the bounding
trapezoid to decide the boundaries of the iteration through the map,
rather than iterating every tile and checking mins/maxs inside the
loop.

Finally, store motion parameters as vectors in chopper and missile
code so that interpolation doesn't require any trig functions but also
so the code is generic enough that we'll rip it out in to core physics
code later.