Implement entity base class for missile and chopper
This is just the basic framework for the API but eventually this will
be used to:
 1. Consolidate some physics and motion interpolation code
 2. Unify some of the rendering path of chopper, missiles and other items
 3. Implement culling for items and missiles
 4. Allow us to implement a unified collision detection

This patch also gets rid of the swappable render backends for the texture
subsystem. The only render backend we're supporting is OpenGL now so it's
pointless to pass renderer_t's around all over the place just to load
a font or texture.